requirements.txt
django>=1.11,<1.12
django-debug-toolbar==1.8
gunicorn==19.4.5
psycopg2==2.7.3.1
whitenoise==3.3.1
Pillow==5.2.0
django-ckeditor==5.6.1
django-bootstrap4==0.0.7
djangorestframework==3.9.0
django-taggit==0.23.0
django-bootstrap-form==3.4
django-filter==2.0.0
